+++ trunk/src/misc/vector.cpp
@@ -7,7 +7,7 @@ void Vector2D::unit()
 {
     if (!isNull()) // Si le vecteur n'est pas nul
     {
-        float n = 1.0f / sqrt(x*x + y*y);    // Inverse de la norme du vecteur
+        float n = 1.0f / sqrtf(x*x + y*y);    // Inverse de la norme du vecteur
         x *= n;
         y *= n;
     }
